requirements.txt
anaconda==0.0.1.1
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.4
oauthlib==3.1.0
pandas==0.24.2
Pillow==6.2.1
pygame==1.9.6
pylint==2.3.1
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.8.0
typed-ast==1.3.5
urllib3==1.25.3
Werkzeug==0.16.0
wrapt==1.11.1
